pyproject.toml
[project]
name = "fastrtc-groq"
version = "0.1.0"
description = "A simple voice agent using FastRTC and Groq"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastrtc[vad]>=0.0.19",
    "groq>=0.22.0",
    "numpy>=2.1.3",
    "langgraph>=0.1.18",
    "langchain-core>=0.1.29",
    "langchain-groq>=0.1.5",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.1",
    "kokoro-onnx>=0.4.9",
]
